rename email.py to avoid clashes with standard library module "email"
